requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in
#
asttokens==2.0.8
    # via icontract
cachetools==4.2.4
    # via google-auth
certifi==2022.9.14
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.0.3
    # via -r requirements/requirements.in
codetiming==1.3.0
    # via -r requirements/requirements.in
google-api-core==1.31.6
    # via
    #   -r requirements/requirements.in
    #   google-api-python-client
    #   google-cloud-bigquery
    #   google-cloud-core
google-api-python-client==1.12.11
    # via -r requirements/requirements.in
google-auth==1.35.0
    # via
    #   -r requirements/requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-bigquery==1.28.0
    # via -r requirements/requirements.in
google-cloud-core==1.7.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-storage==1.35.1
    # via -r requirements/requirements.in
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==1.3.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.56.4
    # via google-api-core
httplib2==0.20.4
    # via
    #   google-api-python-client
    #   google-auth-httplib2
humanfriendly==10.0
    # via -r requirements/requirements.in
icontract==2.6.1
    # via -r requirements/requirements.in
idna==3.4
    # via requests
packaging==21.3
    # via google-api-core
protobuf==3.17.3
    # via
    #   -r requirements/requirements.in
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.9
    # via
    #   httplib2
    #   packaging
pytz==2022.2.1
    # via google-api-core
pyyaml==5.4.1
    # via -r requirements/requirements.in
requests==2.28.1
    # via
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   asttokens
    #   google-api-core
    #   google-api-python-client
    #   google-auth
    #   google-auth-httplib2
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-resumable-media
    #   protobuf
    #   tenacity
tenacity==7.0.0
    # via -r requirements/requirements.in
typing-extensions==4.3.0
    # via icontract
uritemplate==3.0.1
    # via google-api-python-client
urllib3==1.26.12
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
